code/requirements.txt
# LangChain 核心库
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# 向量数据库
lancedb>=0.4.0

# 机器学习框架
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# 图数据库和网络分析
networkx>=3.0
neo4j>=5.0.0

# 其他工具
tavily-python>=0.3.0

# 基础依赖
python-dotenv>=1.0.0
